requirements.txt
pip-tools
justai
pillow
anthropic>=0.62.0
pyyaml